Refactor character_counts_to_string
This commit includes a couple updates to the character_counts_to_string
method. Specifically, the roman_character_index enum is now used in the
main loop instead of an integer, and the variable storing the string
output of the method is renamed from roman_numeral to result.

A minor update has also been made to the .gitignore to ignore Emacs
backup files.