Passing test L - I == XLIX
Note we are adopting the rather common modern convention that
subtractive forms AB (with for "Roman digits" A and B with A < B) are
not permitted unless there is at most one Roman digit between A and
B. An easy modification would permit this to be changed during
compilation via a #define declaration and minor tweaks to our
subtractive form flagging function. Since this behavior isn't requested
in the kata, however, we opt for the simpler implementation.