Pass test MMD + MMD == MMMMM
The kata instructions are ambiguous regarding whether or not to accept
input exceeding 3999 (the largest integer one can express as a Roman
numeral without using four or more 'M' characters). With this test we
fix a bug that yielded incorrect but non-failing output when the sum was
greater than 3999. Now the correct output is produced, with more than
three copies of M possibly in the output string. By design (and as can
be seen in the size of the roman_calculator.c diff), removing support
for output greater than 3999 could easily be implemented by exiting
early in compute_carryovers (or in the add_roman_numerals after
carryovers as a separate check).